requirements.txt
# Essential ML packages for Gymbite nutrition model
numpy>=1.21.0
pandas>=1.3.0
scikit-learn==1.7.0
matplotlib>=3.5.0
joblib>=1.1.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
